requirements.txt
streamlit
sqlalchemy
psycopg2-binary
python-dotenv
pandas
numpy
matplotlib
seaborn
statsmodels
scipy
scikit-learn
